moved the mongodb queries to task_run_mongo_util.py
